+++ lib/yield_star_client/floor_plan_methods.rb
@@ -44,44 +44,36 @@ module YieldStarClient
     end
 
     def get_floor_plans_with_units(external_property_id)
-      floor_plans = get_floor_plans(external_property_id)
-      available_floor_plans = get_available_floor_plans(external_property_id)
+      request_args = default_savon_params.merge(
+        external_property_id: external_property_id
+      )
 
-      names = floor_plans_without_available_units(floor_plans, available_floor_plans)
-      available_floor_plans + floor_plan_with_names(floor_plans, names)
+      response = GetRentSummary::Request.execute(request_args)
+
+      floor_plans = GetRentSummary::Response.new(response)
+        .rent_summaries_as_floor_plans
+
+      available_units = get_available_units(external_property_id)
+
+      floor_plans.map do |floor_plan|
+        floor_plan.units = available_units.select do |unit|
+          unit.unit_type == floor_plan.id
+        end
+      end.flatten
+
+      floor_plans
     end
 
     private
 
-    def floor_plan_with_names(floor_plans, names)
-      floor_plans.select do |floor_plan|
-        names.include? floor_plan.name
-      end
-    end
-
-    def floor_plans_without_available_units(floor_plans, available_floor_plans)
-      floor_plans.map(&:name) - available_floor_plans.map(&:name)
-    end
-
-    # Retrieves rental information for all currently available units at a specific
-    # property, grouped by floor plan.
-    #
-    # @param [String] external_property_id the ID of the property where the available
-    #                                      units are located
-    #
-    # @return [Array<YieldStarClient::AvailableFloorPlan>] list of floor plans with
-    #                                                      available units
-    #
-    # @raise [ArgumentError] when a required argument is missing or invalid
-    # @raise [YieldStarClient::AuthenticationError] when unable to authenticate to the web service
-    # @raise [YieldStarClient::OperationError] when the service raises an OperationError fault
-    # @raise [YieldStarClient::InternalError] when the service raises an InternalError fault
-    # @raise [YieldStarClient::ServerError] when any other server-side error occurs
-    def get_available_floor_plans(external_property_id)
+    def get_available_units(external_property_id)
       request_args = default_savon_params.
         merge(external_property_id: external_property_id)
+
       response = GetAvailableUnits::Request.execute(request_args)
-      GetAvailableUnits::Response.new(response).available_floor_plans
+
+      GetAvailableUnits::Response.new(response)
+        .available_units
     end
   end
 end