+++ lib/yield_star_client/models/floor_plan.rb
@@ -1,35 +1,39 @@
 module YieldStarClient
+  class FloorPlan
 
-  # Represents a floor plan in the YieldStar system.
-  #
-  # A floor plan is guaranteed to have an +external_property_id+ and +name+. All other attributes are
-  # optional.
-  #
-  # @attr [String] external_property_id the property ID of this floor plan
-  # @attr [String] name the name of this floor plan
-  # @attr [String] description the description of this floor plan
-  # @attr [Integer] square_feet the average square footage of this floor plan
-  # @attr [Integer] unit_count the number of units with this floor plan
-  # @attr [Float] bedrooms the bedroom count of the floor plan
-  # @attr [Float] bathrooms the bathroom count of the floor plan
-  class FloorPlan < Modelish::Base
-    property :external_property_id
-    property :name
-    property :square_feet, :type => Integer
-    property :unit_count, :type => Integer
-    property :bedrooms, :type => Float
-    property :bathrooms, :type => Float
-    property :units, :from => :unit, :default => []
+    include ::Virtus.model
 
-    def initialize(options = {})
-      super(options)
+    attribute :external_property_id, String
+    attribute :effective_date, Date
+    attribute :floor_plan_name, String
+    attribute :bedrooms, Float
+    attribute :bathrooms, Float
+    attribute :square_feet, Integer
+    attribute :units, Array
+    attribute :id, String
+    attribute :name, String
+    attribute :min_market_rent, Integer
+    attribute :max_market_rent, Integer
+    attribute :min_final_rent, Integer
+    attribute :max_final_rent, Integer
 
-      self.units = [self.units].flatten.collect { |u| AvailableUnit.new(u) }
-    end
+    def self.new_from_rent_summary_hash(hash)
+      floorplan_id = hash.delete(:unit_type)
+      square_feet = hash.delete(:avg_sq_ft)
+      bedrooms = hash.delete(:bed_rooms)
+      bathrooms = hash.delete(:bath_rooms)
 
-    def id
-      name
+      new(
+        hash.merge(
+          id: floorplan_id,
+          name: floorplan_id,
+          square_feet: square_feet,
+          bedrooms: bedrooms,
+          bathrooms: bathrooms
+        )
+      )
     end
 
   end
+
 end