+++ spec/yield_star_client/lease_term_rent_methods_spec.rb
@@ -10,7 +10,7 @@ module YieldStarClient
       ))
     end
 
-    describe "#get_lease_term_rent", vcr: {record: :once} do
+    describe "#get_lease_term_rent", vcr: { record: :once } do
       let(:properties) { client.get_properties }
       let(:external_property_id) { properties.last.external_property_id }
       let(:floor_plan) do
@@ -44,8 +44,6 @@ module YieldStarClient
 
         expect(lease_term_rents).to_not be_empty
         expect(lease_term_rents.first).to be_a LeaseTermRent
-        expect(lease_term_rents.map(&:unit_number).uniq).to eq ["1904", "0509"]
-        expect(lease_term_rents.map(&:building).uniq).to eq ["Z19", "Z05"]
       end
     end